feat: wire tonic dopamine posture into consciousness daemon gap-picker
- run_cycle reads tonic before gap scan: low DA parks voids in consolidation mode
- picker filters cooldown-active nodes before selection and rotates across fresh targets (epoch-slot modulo) instead of always attacking the largest pocket
- gap discovery fires boost_tonic(+0.04), closing the arousal loop
- fixes stamp_bonus type contract (phasic: float)

+++ dopamine_state.py
@@ -145,7 +145,7 @@ def decay_modulator(tonic: float) -> float:
     return DECAY_MULT_CEIL + (DECAY_MULT_FLOOR - DECAY_MULT_CEIL) * t
 
 
-def stamp_bonus(phasic: int) -> int:
+def stamp_bonus(phasic: float) -> int:
     """
     Importance points added by DeepMemory.store() for memories born during
     a phasic spike. 0 at rest, up to +2 at full spike.
